perf(canvas): lazy-load plugin market runtime
