+++ apps/canvas-workspace/src/main/settings/canvas-plugins-config.ts
@@ -9,9 +9,7 @@ import type {
   CanvasPluginsImportEntry,
   CanvasPluginsStatus,
 } from '../../shared/settings-config';
-import { readPluginPackage } from '../plugin-market/package-reader';
 import { agentPluginSkillScanPathsSync } from '../plugin-market/skill-scan';
-import { canvasEntryFromPackage, dedupeRendererSpecs } from '../plugin-market/canvas-package-adapter';
 
 interface CanvasPluginsConfigFile {
   pluginDirs?: string[];
@@ -242,6 +240,10 @@ async function readPluginEntry(
   dir: string,
   config: CanvasPluginsConfigFile,
 ): Promise<CanvasPluginEntry> {
+  const [{ readPluginPackage }, { canvasEntryFromPackage }] = await Promise.all([
+    import('../plugin-market/package-reader'),
+    import('../plugin-market/canvas-package-adapter'),
+  ]);
   const result = await readPluginPackage(dir);
   if (!result.package) {
     return {
@@ -323,6 +325,7 @@ export async function getCanvasPluginsStatus(): Promise<CanvasPluginsStatus> {
   const config = await readConfig();
   const pluginDirs = Array.from(new Set((config.pluginDirs ?? []).map(normalizePluginDir).filter(Boolean)));
   const plugins = await Promise.all(pluginDirs.map((dir) => readPluginEntry(dir, config)));
+  const { dedupeRendererSpecs } = await import('../plugin-market/canvas-package-adapter');
   return {
     path: canvasPluginsConfigPath(),
     pluginDirs,
